pyproject.toml
[tool.poetry]
name = "noneflow"
version = "2.7.3"
description = "Manage publish related issues in nonebot2 project"
authors = ["hemengyang <hmy0119@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/nonebot/noneflow"
repository = "https://github.com/nonebot/noneflow"

[tool.poetry.dependencies]
python = "^3.10"
nonebot2 = { extras = ["httpx"], version = "^2.0.0" }
nonebot-adapter-github = "^0.3.0"
pre-commit = "^3.3.2"
jinja2 = "^3.1.2"

[tool.poetry.group.plugin.dependencies]
click = "^8.1.3"
tzdata = "^2023.3"

[tool.poetry.group.dev.dependencies]
isort = "*"
black = "*"
nonebug = "^0.3.2"
pytest-cov = "^4.0.0"
pytest-mock = "^3.6.1"
pytest-asyncio = "^0.21.0"
respx = "^0.20.1"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88
skip_gitignore = true

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.nonebot]
adapters = [{ name = "GitHub", module_name = "nonebot.adapters.github" }]
plugin_dirs = ["src/plugins"]

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
